Revert "Fixed: Breaking trakt API changes (Posters for add series are placeholders for now)"
This reverts commit 66ca925b0d4fcd5f7085795b757d8f63b0c1e9a7.

+++ src/NzbDrone.Core/MetadataSource/Trakt/Ratings.cs
@@ -4,7 +4,7 @@ namespace NzbDrone.Core.MetadataSource.Trakt
 {
     public class Ratings
     {
-        public Decimal percentage { get; set; }
+        public Int32 percentage { get; set; }
         public Int32 votes { get; set; }
         public Int32 loved { get; set; }
         public Int32 hated { get; set; }
